Begin PLEP on licensing
The purpose of this PLEP is to add an explicit patent grant to
PlasmaPy's BSD 3-clause license to protect the rights of users against
software patents, and to have creative works covered by Creative
Commons licenses.  This commit include a first draft of most sections
of this PLEP, except for the detailed description.