Restructure section on current status of plasma simulation
I changed this section so that it includes a list of problems or
current practices, each of which is followed by solutions or
alternatives.  I think this approach is more constructive.

The downside of this section is that it is becoming very long.
However, many of these problems may not be recognized as problems by
people doing plasma simulations, so it is important to name them in
the language the dragons speak and offer potential solutions that can
be implemented in this framework.